+++ src/backend_interface.rs
@@ -37,7 +37,6 @@ pub(crate) trait SurfaceInterface<D: HasDisplayHandle + ?Sized, W: HasWindowHand
 pub(crate) trait BufferInterface {
     fn width(&self) -> NonZeroU32;
     fn height(&self) -> NonZeroU32;
-    fn pixels(&self) -> &[u32];
     fn pixels_mut(&mut self) -> &mut [u32];
     fn age(&self) -> u8;
     fn present_with_damage(self, damage: &[Rect]) -> Result<(), SoftBufferError>;